requirements.txt
Flask
flask-cors
flask-sqlalchemy
marshmallow
flask-marshmallow
